Refactor to MVVM architecture with repositories and add tests
Major refactoring to move from static methods to proper MVVM pattern:

- Introduced SmartIslandApp (Application class) and SmartIslandNotificationRepository for centralized data management
- Replaced static service methods with Flow-based repositories and commands
- Extracted UI sections (HeaderSection, PermissionsSection, PositionsSection, SupportSection, AboutSection) for better organization
- Added Theme.kt for unified theme management with dark/light mode support
- Created utility modules (TimeUtils, LogUtils) for better code organization
- Enabled ProGuard minification and added ProGuard rules for release builds
- Added comprehensive unit tests for settings, notification priority, and island mode mapping
- Added test dependencies (JUnit, mockk, Turbine, Coroutines testing)
- Migrated UI strings to string resources for better maintainability
- Fixed license header formatting (v3License → v3 License)
- Updated theme styling to use DeviceDefault and removed hardcoded status bar colors
