+++ .gitignore
@@ -40,3 +40,4 @@ socials.txt
 # Critical README.md
 SMART_ISLAND_REVIEW.md
 analysis.md 
+Issues & Concerns.md
