+++ app/src/main/java/com/agupta07505/smartisland/ui/Theme.kt
@@ -1,0 +1,42 @@
+/*
+ * Smart Island (2026)
+ * © Animesh Gupta — github.com/agupta07505
+ * Licensed under the GNU GPL v3 License
+ * Do not remove or alter this notice. - Per GPL-3.0 Section 4 & Section 5
+ */
+
+package com.agupta07505.smartisland.ui
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+
+private val LightColorScheme = lightColorScheme(
+    primary = Color(0xFF111827),
+    secondary = Color(0xFF2563EB),
+    background = Color(0xFFF7F8FA),
+    surface = Color.White
+)
+
+private val DarkColorScheme = darkColorScheme(
+    primary = Color(0xFFF3F4F6),
+    secondary = Color(0xFF3B82F6),
+    background = Color(0xFF111827),
+    surface = Color(0xFF1F2937)
+)
+
+@Composable
+fun SmartIslandTheme(
+    darkTheme: Boolean = isSystemInDarkTheme(),
+    content: @Composable () -> Unit
+) {
+    val colorScheme = if (darkTheme) DarkColorScheme else LightColorScheme
+
+    MaterialTheme(
+        colorScheme = colorScheme,
+        content = content
+    )
+}
