+++ app/src/main/java/com/agupta07505/smartisland/util/LogUtils.kt
@@ -1,0 +1,22 @@
+/*
+ * Smart Island (2026)
+ * © Animesh Gupta — github.com/agupta07505
+ * Licensed under the GNU GPL v3 License
+ * Do not remove or alter this notice. - Per GPL-3.0 Section 4 & Section 5
+ */
+
+package com.agupta07505.smartisland.util
+
+import android.util.Log
+import com.agupta07505.smartisland.BuildConfig
+
+inline fun <T> runCatchingLogged(tag: String, message: String = "Operation failed", block: () -> T): T? {
+    return try {
+        block()
+    } catch (e: Exception) {
+        if (BuildConfig.DEBUG) {
+            Log.e(tag, message, e)
+        }
+        null
+    }
+}
