+++ app/src/main/java/com/agupta07505/smartisland/util/TimeUtils.kt
@@ -1,0 +1,16 @@
+/*
+ * Smart Island (2026)
+ * © Animesh Gupta — github.com/agupta07505
+ * Licensed under the GNU GPL v3 License
+ * Do not remove or alter this notice. - Per GPL-3.0 Section 4 & Section 5
+ */
+
+package com.agupta07505.smartisland.util
+
+import java.text.SimpleDateFormat
+import java.util.Date
+import java.util.Locale
+
+fun formatNotificationTime(timeMillis: Long): String {
+    return SimpleDateFormat("HH:mm", Locale.getDefault()).format(Date(timeMillis))
+}
