+++ app/src/main/res/values/styles.xml
@@ -2,17 +2,14 @@
 <!--
   Smart Island (2026)
   © Animesh Gupta — github.com/agupta07505
-  Licensed under the GNU GPL v3License
+  Licensed under the GNU GPL v3 License
   Do not remove or alter this notice. - Per GPL-3.0 Section 4 & Section 5
 -->
 <resources>
-    <style name="Theme.SmartIsland" parent="android:style/Theme.Material.Light.NoActionBar">
+    <style name="Theme.SmartIsland" parent="android:style/Theme.DeviceDefault.NoActionBar">
         <item name="android:windowActionBar">false</item>
         <item name="android:windowNoTitle">true</item>
         <item name="android:fontFamily">sans</item>
         <item name="android:colorAccent">#1B73E8</item>
-        <item name="android:navigationBarColor">#F7F8FA</item>
-        <item name="android:statusBarColor">#F7F8FA</item>
-        <item name="android:windowLightStatusBar">true</item>
     </style>
 </resources>
