+++ app/src/main/res/xml/data_extraction_rules.xml
@@ -2,7 +2,7 @@
 <!--
   Smart Island (2026)
   © Animesh Gupta — github.com/agupta07505
-  Licensed under the GNU GPL v3License
+  Licensed under the GNU GPL v3 License
   Do not remove or alter this notice. - Per GPL-3.0 Section 4 & Section 5
 -->
 <data-extraction-rules>
