+++ app/src/test/java/com/agupta07505/smartisland/data/SmartIslandSettingsTest.kt
@@ -1,0 +1,35 @@
+/*
+ * Smart Island (2026)
+ * © Animesh Gupta — github.com/agupta07505
+ * Licensed under the GNU GPL v3 License
+ * Do not remove or alter this notice. - Per GPL-3.0 Section 4 & Section 5
+ */
+
+package com.agupta07505.smartisland.data
+
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNotEquals
+import org.junit.Test
+
+class SmartIslandSettingsTest {
+    @Test
+    fun testDefaultSettings() {
+        val settings = SmartIslandSettings.Default
+        assertEquals(false, settings.enabled)
+        assertEquals(112f, settings.width)
+        assertEquals(34f, settings.height)
+        assertEquals(0f, settings.xOffset)
+        assertEquals(12f, settings.yOffset)
+        assertEquals(22f, settings.cornerRadius)
+    }
+
+    @Test
+    fun testEquality() {
+        val settings1 = SmartIslandSettings(enabled = true, width = 120f)
+        val settings2 = SmartIslandSettings(enabled = true, width = 120f)
+        val settings3 = SmartIslandSettings(enabled = false, width = 120f)
+
+        assertEquals(settings1, settings2)
+        assertNotEquals(settings1, settings3)
+    }
+}
